+++ bin/apis/linetv-list-modules-request.js
@@ -16,7 +16,7 @@ class LINETvListModulesRequest extends _linetvRequest.default {
   }
   /**
     * @param {number} channelId
-    * @param {string} channelCode
+    * @param {string} countryCode
     * @return {AxiosResponse<LINETvListModulesResponseData>}
     */