+++ bin/typedef/linetv-get-station-response-data.js
@@ -1,0 +1,7 @@
+/**
+ * @typedef LINETvGetStationResponseData
+ * @property {{ message: string; code: number }} header
+ * @property {LINETvStationData} body
+ */
+"use strict";
+//# sourceMappingURL=linetv-get-station-response-data.js.map