+++ bin/typedef/linetv-live-data.js
@@ -1,0 +1,15 @@
+/**
+ * @typedef LINETvLiveData
+ * @property {number} serviceUrl
+ * @property {string} channelName
+ * @property {number} liveNo
+ * @property {number} playCount
+ * @property {string} thumbnailUrl
+ * @property {string} liveStatus
+ * @property {string} liveStartDate
+ * @property {string} liveEndDate
+ * @property {string} liveTitle
+ * @property {number} likeitPoint
+ */
+"use strict";
+//# sourceMappingURL=linetv-live-data.js.map