+++ bin/typedef/linetv-ranking-response-data.js
@@ -1,0 +1,7 @@
+/**
+ * @typedef LINETvRankingResponseData
+ * @property {{ message: string; code: number }} header
+ * @property {{ clips: Array<LINETvClipData> ; hasMore: boolean}} body
+ */
+"use strict";
+//# sourceMappingURL=linetv-ranking-response-data.js.map